Django_new/django-fyp-new-1st-may/requirements.txt
Django
djangorestframework
python-dotenv
google-generativeai
langchain
langchain-community
langchain-core
langchain-google-genai
huggingface-hub
transformers
reportlab
faiss-cpu
pypdf
PyMuPDF
pypdfium2
requests
dotenv
sentence-transformers